always use unsigned short vars/vals with "%hu" format spec
git-svn-id: svn://tug.org/texlive/trunk/Build/source@18813 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/dvipdfmx/src/dpxfile.c
@@ -726,7 +726,7 @@ if ((l) + (n) >= (m)) { \
         }
       case  'v': /* Version number, e.g. 1.4 */ {
        char buf[6];
-       sprintf(buf, "1.%hu", version);
+       sprintf(buf, "1.%hu", (unsigned short) version);
        need(cmd, n, size, strlen(buf));
        strcpy(cmd + n, buf);  n += strlen(buf);
        break;
